Don't exit monitoring when lag data error is encountered (#7)
The monitoring refresh loop normally exits on any error. In testing
it seems like the Spock lag data is sometimes not available which
leads to an error. This change alters the monitoring command so
it will not exit and will attempt to pull the lag values in the
next refresh.